+++ WrightTools/data/__init__.py
@@ -3,7 +3,6 @@
 
 
 from ._brunold import *
-from ._cary import *
 from ._colors import *
 from ._data import *
 from ._jasco import *
