+++ tests/data/from_PyCMDS.py
@@ -18,6 +18,7 @@ def test_wm_w2_w1_000():
     data = wt.data.from_PyCMDS(p)
     assert data.shape == (35, 11, 11)
     assert data.axis_names == ['wm', 'w2', 'w1']
+    data.close()
 
 
 def test_wm_w2_w1_001():
@@ -25,3 +26,4 @@ def test_wm_w2_w1_001():
     data = wt.data.from_PyCMDS(p)
     assert data.shape == (29, 11, 11)
     assert data.axis_names == ['wm', 'w2', 'w1']
+    data.close()
